Split developer settings into 2 screens to be able to access global settings when no logged in.
