+++ features/preferences/impl/src/main/kotlin/io/element/android/features/preferences/impl/developer/DeveloperSettingsEvents.kt
@@ -9,15 +9,8 @@
 package io.element.android.features.preferences.impl.developer
 
 import androidx.compose.ui.graphics.Color
-import io.element.android.features.preferences.impl.developer.tracing.LogLevelItem
-import io.element.android.libraries.featureflag.ui.model.FeatureUiModel
-import io.element.android.libraries.matrix.api.tracing.TraceLogPack
 
 sealed interface DeveloperSettingsEvents {
-    data class UpdateEnabledFeature(val feature: FeatureUiModel, val isEnabled: Boolean) : DeveloperSettingsEvents
-    data class SetCustomElementCallBaseUrl(val baseUrl: String?) : DeveloperSettingsEvents
-    data class SetTracingLogLevel(val logLevel: LogLevelItem) : DeveloperSettingsEvents
-    data class ToggleTracingLogPack(val logPack: TraceLogPack, val enabled: Boolean) : DeveloperSettingsEvents
     data class SetShowColorPicker(val show: Boolean) : DeveloperSettingsEvents
     data class ChangeBrandColor(val color: Color?) : DeveloperSettingsEvents
     data object ClearCache : DeveloperSettingsEvents
