+++ features/preferences/impl/src/main/kotlin/io/element/android/features/preferences/impl/developer/DeveloperSettingsState.kt
@@ -8,32 +8,19 @@
 
 package io.element.android.features.preferences.impl.developer
 
-import io.element.android.features.preferences.impl.developer.tracing.LogLevelItem
-import io.element.android.features.rageshake.api.preferences.RageshakePreferencesState
+import io.element.android.features.preferences.impl.developer.appsettings.AppDeveloperSettingsState
 import io.element.android.libraries.architecture.AsyncAction
 import io.element.android.libraries.architecture.AsyncData
-import io.element.android.libraries.featureflag.ui.model.FeatureUiModel
-import io.element.android.libraries.matrix.api.tracing.TraceLogPack
-import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.ImmutableMap
 
 data class DeveloperSettingsState(
-    val features: ImmutableList<FeatureUiModel>,
+    val appDeveloperSettingsState: AppDeveloperSettingsState,
     val cacheSize: AsyncData<String>,
     val databaseSizes: AsyncData<ImmutableMap<String, String>>,
-    val rageshakeState: RageshakePreferencesState,
     val clearCacheAction: AsyncAction<Unit>,
-    val customElementCallBaseUrlState: CustomElementCallBaseUrlState,
-    val tracingLogLevel: AsyncData<LogLevelItem>,
-    val tracingLogPacks: ImmutableList<TraceLogPack>,
     val isEnterpriseBuild: Boolean,
     val showColorPicker: Boolean,
     val eventSink: (DeveloperSettingsEvents) -> Unit
 ) {
     val showLoader = clearCacheAction is AsyncAction.Loading
 }
-
-data class CustomElementCallBaseUrlState(
-    val baseUrl: String?,
-    val validator: (String?) -> Boolean,
-)
