+++ features/preferences/impl/src/main/kotlin/io/element/android/features/preferences/impl/developer/DeveloperSettingsView.kt
@@ -10,40 +10,28 @@ package io.element.android.features.preferences.impl.developer
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.progressSemantics
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import io.element.android.compound.theme.ElementTheme
 import io.element.android.features.preferences.impl.R
-import io.element.android.features.preferences.impl.developer.tracing.LogLevelItem
-import io.element.android.features.rageshake.api.preferences.RageshakePreferencesView
+import io.element.android.features.preferences.impl.developer.appsettings.AppDeveloperSettingsView
 import io.element.android.libraries.designsystem.components.ProgressDialog
 import io.element.android.libraries.designsystem.components.list.ListItemContent
 import io.element.android.libraries.designsystem.components.preferences.PreferenceCategory
-import io.element.android.libraries.designsystem.components.preferences.PreferenceDropdown
 import io.element.android.libraries.designsystem.components.preferences.PreferencePage
-import io.element.android.libraries.designsystem.components.preferences.PreferenceSwitch
-import io.element.android.libraries.designsystem.components.preferences.PreferenceTextField
 import io.element.android.libraries.designsystem.preview.ElementPreview
 import io.element.android.libraries.designsystem.preview.PreviewsDayNight
 import io.element.android.libraries.designsystem.theme.components.CircularProgressIndicator
 import io.element.android.libraries.designsystem.theme.components.ListItem
 import io.element.android.libraries.designsystem.theme.components.Text
-import io.element.android.libraries.featureflag.ui.FeatureListView
-import io.element.android.libraries.featureflag.ui.model.FeatureUiModel
-import io.element.android.libraries.matrix.api.tracing.TraceLogPack
 import io.element.android.libraries.ui.strings.CommonStrings
 import io.mhssn.colorpicker.ColorPickerDialog
 import io.mhssn.colorpicker.ColorPickerType
-import kotlinx.collections.immutable.toImmutableList
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
@@ -71,52 +59,12 @@ fun DeveloperSettingsView(
         title = stringResource(id = CommonStrings.common_developer_options)
     ) {
         // Note: this is OK to hardcode strings in this debug screen.
-        PreferenceCategory(
-            title = "Feature flags",
-        ) {
-            FeatureListContent(state)
-        }
-        NotificationCategory(onPushHistoryClick)
-        ElementCallCategory(state = state)
-
-        PreferenceCategory(title = "Rust SDK") {
-            PreferenceDropdown(
-                title = "Tracing log level",
-                supportingText = "Requires app reboot",
-                selectedOption = state.tracingLogLevel.dataOrNull(),
-                options = LogLevelItem.entries.toImmutableList(),
-                onSelectOption = { logLevel ->
-                    state.eventSink(DeveloperSettingsEvents.SetTracingLogLevel(logLevel))
-                }
-            )
-        }
-        PreferenceCategory(title = "Enable trace logs per SDK feature") {
-            Text(
-                text = "Requires app reboot",
-                style = ElementTheme.typography.fontBodyMdRegular,
-                color = ElementTheme.colors.textSecondary,
-                modifier = Modifier.padding(start = 16.dp, end = 16.dp, bottom = 8.dp)
-            )
-            for (logPack in TraceLogPack.entries) {
-                PreferenceSwitch(
-                    title = logPack.title,
-                    isChecked = state.tracingLogPacks.contains(logPack),
-                    onCheckedChange = { isChecked -> state.eventSink(DeveloperSettingsEvents.ToggleTracingLogPack(logPack, isChecked)) }
-                )
-            }
-        }
-
-        PreferenceCategory(title = "Showkase") {
-            ListItem(
-                headlineContent = {
-                    Text("Open Showkase browser")
-                },
-                onClick = onOpenShowkase
-            )
-        }
-        RageshakePreferencesView(
-            state = state.rageshakeState,
+        AppDeveloperSettingsView(
+            state = state.appDeveloperSettingsState,
+            onOpenShowkase = onOpenShowkase,
         )
+        NotificationCategory(onPushHistoryClick)
+
         if (state.isEnterpriseBuild) {
             PreferenceCategory(title = "Theme") {
                 ListItem(
@@ -136,14 +84,6 @@ fun DeveloperSettingsView(
                     }
                 )
             }
-        }
-        PreferenceCategory(title = "Crash") {
-            ListItem(
-                headlineContent = {
-                    Text("Crash the app 💥")
-                },
-                onClick = { error("This crash is a test.") }
-            )
         }
         val cache = state.cacheSize
         PreferenceCategory(title = "Cache") {
@@ -213,32 +153,6 @@ fun DeveloperSettingsView(
 }
 
 @Composable
-private fun ElementCallCategory(
-    state: DeveloperSettingsState,
-) {
-    PreferenceCategory(title = "Element Call") {
-        val callUrlState = state.customElementCallBaseUrlState
-
-        val supportingText = if (callUrlState.baseUrl.isNullOrEmpty()) {
-            stringResource(R.string.screen_advanced_settings_element_call_base_url_description)
-        } else {
-            callUrlState.baseUrl
-        }
-        PreferenceTextField(
-            headline = stringResource(R.string.screen_advanced_settings_element_call_base_url),
-            value = callUrlState.baseUrl,
-            placeholder = "https://.../room",
-            supportingText = supportingText,
-            validation = callUrlState.validator,
-            onValidationErrorMessage = stringResource(R.string.screen_advanced_settings_element_call_base_url_validation_error),
-            displayValue = { value -> !value.isNullOrEmpty() },
-            keyboardOptions = KeyboardOptions.Default.copy(autoCorrectEnabled = false, keyboardType = KeyboardType.Uri),
-            onChange = { state.eventSink(DeveloperSettingsEvents.SetCustomElementCallBaseUrl(it)) }
-        )
-    }
-}
-
-@Composable
 private fun NotificationCategory(onPushHistoryClick: () -> Unit) {
     PreferenceCategory(title = stringResource(id = R.string.screen_notification_settings_title)) {
         ListItem(
@@ -250,20 +164,6 @@ private fun NotificationCategory(onPushHistoryClick: () -> Unit) {
     }
 }
 
-@Composable
-private fun FeatureListContent(
-    state: DeveloperSettingsState,
-) {
-    fun onFeatureEnabled(feature: FeatureUiModel, isEnabled: Boolean) {
-        state.eventSink(DeveloperSettingsEvents.UpdateEnabledFeature(feature, isEnabled))
-    }
-
-    FeatureListView(
-        features = state.features,
-        onCheckedChange = ::onFeatureEnabled,
-    )
-}
-
 @PreviewsDayNight
 @Composable
 internal fun DeveloperSettingsViewPreview(
@@ -273,6 +173,6 @@ internal fun DeveloperSettingsViewPreview(
         state = state,
         onOpenShowkase = {},
         onPushHistoryClick = {},
-        onBackClick = {}
+        onBackClick = {},
     )
 }
